Made second logo span the viewport width better. 🤔
I figured that two units of padding was unnecessarily too much, and decided to reduce it to only one.

Notice, however, that it’s important for there to be padding because of the overshoot of the “g” and the “d”.

The horizontal padding is not required, but I decided to keep it to avoid a half‐unit, as well as to balance out with the vertical padding.

Herein I also decided to take the opportunity to change the ‘clip-path’ of the ‘a’s to be a little cleaner. It probably will work on fewer SVG renderers, but oh well. *Shrugs*

Notice that I have only tested for this change to work on Firefox (v63.0). As a todo, I should probably test it on Chrome too once I have access to my computer again.